+++ ddtrace/contrib/asgi/middleware.py
@@ -72,6 +72,11 @@ def _extract_headers(scope):
     return {}
 
 
+def _default_handle_exception_span(exc, span):
+    """Default handler for exception for span"""
+    span.set_tag(http.STATUS_CODE, 500)
+
+
 class TraceMiddleware:
     """
     ASGI application middleware that traces the requests.
@@ -81,10 +86,11 @@ class TraceMiddleware:
         tracer: Custom tracer. Defaults to the global tracer.
     """
 
-    def __init__(self, app, tracer=None, config=config.asgi):
+    def __init__(self, app, tracer=None, config=config.asgi, handle_exception_span=_default_handle_exception_span):
         self.app = guarantee_single_callable(app)
         self.tracer = tracer or ddtrace.tracer
         self.config = config
+        self.handle_exception_span = handle_exception_span
 
     async def __call__(self, scope, receive, send):
         if scope["type"] != "http":
@@ -127,10 +133,10 @@ class TraceMiddleware:
 
         try:
             return await self.app(scope, receive, wrapped_send)
-        except Exception:
-            span.set_tag(http.STATUS_CODE, 500)
+        except Exception as exc:
             (exc_type, exc_val, exc_tb) = sys.exc_info()
             span.set_exc_info(exc_type, exc_val, exc_tb)
+            self.handle_exception_span(exc, span)
             reraise(exc_type, exc_val, exc_tb)
         finally:
             span.finish()