+++ ddtrace/contrib/sanic/patch.py
@@ -40,10 +40,18 @@ def _wrap_response_callback(span, callback):
     # based on response and finish span before returning response
 
     def update_span(response):
-        span.set_tag(http.STATUS_CODE, response.status)
-        if 500 <= response.status < 600:
-            span.error = 1
-        store_response_headers(response.headers, span, config.sanic)
+        error = 0
+        if isinstance(response, sanic.response.BaseHTTPResponse):
+            status_code = response.status
+            if 500 <= response.status < 600:
+                error = 1
+            store_response_headers(response.headers, span, config.sanic)
+        else:
+            # invalid response causes ServerError exception which must be handled
+            status_code = 500
+            error = 1
+        span.set_tag(http.STATUS_CODE, status_code)
+        span.error = error
         span.finish()
 
     @wrapt.function_wrapper
@@ -84,7 +92,7 @@ def unpatch():
     setattr(sanic, "__datadog_patch", False)
 
 
-def patch_handle_request(wrapped, instance, args, kwargs):
+async def patch_handle_request(wrapped, instance, args, kwargs):
     """Wrapper for Sanic.handle_request"""
     request = kwargs.get("request", args[0])
     write_callback = kwargs.get("write_callback", args[1])
@@ -120,4 +128,4 @@ def patch_handle_request(wrapped, instance, args, kwargs):
     if stream_callback is not None:
         stream_callback = _wrap_response_callback(span, stream_callback)
 
-    return wrapped(request, write_callback, stream_callback, **kwargs)
+    return await wrapped(request, write_callback, stream_callback, **kwargs)