+++ ddtrace/contrib/vertica/__init__.py
@@ -42,10 +42,10 @@ To configure the Vertica integration on an instance-per-instance basis use the
 from ...utils.importlib import require_modules
 
 
-required_modules = ['vertica_python']
+required_modules = ["vertica_python"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
         from .patch import patch, unpatch
 
-        __all__ = [patch, unpatch]
+        __all__ = ["patch", "unpatch"]