+++ tests/__init__.py
@@ -1,10 +1,15 @@
 import contextlib
+import functools
+import inspect
 import os
 import sys
 from contextlib import contextmanager
 
+import pytest
+
 import ddtrace
 from ddtrace import Tracer, Span
+from ddtrace.compat import httplib
 from ddtrace.constants import SPAN_MEASURED_KEY
 from ddtrace.encoding import JSONEncoder
 from ddtrace.ext import http
@@ -790,3 +795,73 @@ def override_global_tracer(tracer):
     ddtrace.tracer = tracer
     yield
     ddtrace.tracer = original_tracer
+
+
+class SnapshotFailed(Exception):
+    pass
+
+
+def snapshot(ignores=None, tracer=ddtrace.tracer):
+    """Performs a snapshot integration test with the testing agent.
+
+    All traces sent to the agent will be recorded and compared to a snapshot
+    created for the test case.
+
+    :param ignores: A list of keys to ignore when comparing snapshots. To refer
+                    to keys in the meta or metrics maps use "meta.key" and
+                    "metrics.key"
+    :param tracer: A tracer providing the agent connection information to use.
+    """
+    ignores = ignores or []
+
+    def dec(f):
+        @functools.wraps(f)
+        def wrapper(*args, **kwargs):
+            if len(args) > 1:
+                self = args[0]
+                clsname = self.__class__.__name__
+            else:
+                clsname = ""
+
+            module = inspect.getmodule(f)
+
+            # Use the fully qualified function name as a unique test token to
+            # identify the snapshot.
+            token = "{}{}{}.{}".format(module.__name__, "." if clsname else "", clsname, f.__name__)
+
+            conn = httplib.HTTPConnection(tracer.writer.api.hostname, tracer.writer.api.port)
+            try:
+                # Signal the start of this test case to the test agent.
+                try:
+                    conn.request("GET", "/test/start?token=%s" % token)
+                except Exception as e:
+                    pytest.fail("Could not connect to test agent: %s" % str(e), pytrace=False)
+
+                r = conn.getresponse()
+                if r.status != 200:
+                    # The test agent returns nice error messages we can forward to the user.
+                    raise SnapshotFailed(r.read().decode())
+
+                # Run the test.
+                ret = f(*args, **kwargs)
+
+                # Flush out any remnant traces.
+                tracer.writer.flush_queue()
+
+                # Query for the results of the test.
+                conn = httplib.HTTPConnection(tracer.writer.api.hostname, tracer.writer.api.port)
+                conn.request("GET", "/test/snapshot?ignores=%s&token=%s" % (",".join(ignores), token))
+                r = conn.getresponse()
+                if r.status != 200:
+                    raise SnapshotFailed(r.read().decode())
+                return ret
+            except SnapshotFailed as e:
+                # Fail the test if a failure has occurred and print out the
+                # message we got from the test agent.
+                pytest.fail(str(e), pytrace=False)
+            finally:
+                conn.close()
+
+        return wrapper
+
+    return dec