+++ tests/contrib/flask/test_request.py
@@ -687,7 +687,7 @@ class FlaskRequestTestCase(BaseFlaskTestCase):
 
         spans = self.get_spans()
 
-        if flask_version >= (0, 12, 0):
+        if flask_version >= (0, 12):
             self.assertEqual(len(spans), 11)
 
             # Assert the order of the spans created
@@ -709,7 +709,6 @@ class FlaskRequestTestCase(BaseFlaskTestCase):
             )
         else:
             self.assertEqual(len(spans), 10)
-
             # Assert the order of the spans created
             self.assertListEqual(
                 [